Hacky fix for color -1

+++ desktop_version/src/Ent.h
@@ -52,7 +52,7 @@ public:
     int flipped = 0;
     int flippedsize = 0;
 
-    Uint32 realcol = 0;
+    long long realcol = 0;
 };
 
 #endif /* ENT_H */
